refactor: Extract routes into their own file
